clean: 属性を表すコンポーネントを分離しました
